+++ src/chrono_gpu/physics/ChSystemGpu.h
@@ -38,8 +38,8 @@ class ChSystemGpuMesh_impl;
 /// Interface to a Chrono::Gpu system.
 class CH_GPU_API ChSystemGpu {
   public:
-    /// Construct system with given sphere radius, density, and big domain dimensions.
-    ChSystemGpu(float sphere_rad, float density, float3 boxDims);
+    /// Construct system with given sphere radius, density, big domain dimensions and center.
+    ChSystemGpu(float sphere_rad, float density, const ChVector<float>& boxDims, ChVector<float> O = ChVector<float>(0));
 
     /// Construct system with a checkpoint file.
     ChSystemGpu(const std::string& checkpoint);
@@ -47,18 +47,12 @@ class CH_GPU_API ChSystemGpu {
     virtual ~ChSystemGpu();
 
     /// Set gravitational acceleration vector.
-    void SetGravitationalAcceleration(const ChVector<float> g);
-    void SetGravitationalAcceleration(const float3 g);
+    void SetGravitationalAcceleration(const ChVector<float>& g);
 
     /// Set particle positions, velocities and angular velocities.
     void SetParticles(const std::vector<ChVector<float>>& points,
                       const std::vector<ChVector<float>>& vels = std::vector<ChVector<float>>(),
                       const std::vector<ChVector<float>>& ang_vels = std::vector<ChVector<float>>());
-    void SetParticlePositions(const std::vector<ChVector<float>>& points,
-                              const std::vector<ChVector<float>>& vels = std::vector<ChVector<float>>(),
-                              const std::vector<ChVector<float>>& ang_vels = std::vector<ChVector<float>>()) {
-        SetParticles(points, vels, ang_vels);
-    }
 
     /// Set particle positions, velocities and angular velocities from a file.
     void ReadParticleFile(const std::string& infilename);
@@ -72,6 +66,11 @@ class CH_GPU_API ChSystemGpu {
     /// Set the big domain to be fixed or not.
     /// If fixed, it will ignore any given position functions.
     void SetBDFixed(bool fixed);
+
+    /// Set the center of the big box domain, relative to the origin of the coordinate system (default: [0,0,0]).
+    /// Note that the domain is always axis-aligned. The user must make sure that all simulation information (particle
+    /// locations, boundaries, meshes...) is consistent with this domain.
+    void SetBDCenter(const ChVector<float>& O);
 
     /// Set flags indicating whether or not a particle is fixed.
     /// MUST be called only once and MUST be called before Initialize.
@@ -217,6 +216,9 @@ class CH_GPU_API ChSystemGpu {
     /// Return particle linear velocity.
     ChVector<float> GetParticleVelocity(int nSphere) const;
 
+    /// Return the total kinetic energy of all particles.
+    float GetParticlesKineticEnergy() const;
+
     /// Return position of BC plane.
     ChVector<float> GetBCPlanePosition(size_t plane_id) const;
 
@@ -295,6 +297,9 @@ class CH_GPU_API ChSystemGpu {
     /// Write contact pair/history to a stream.
     /// WriteCheckpointFile() and WriteContactHistoryFile() are its wrappers.
     void WriteHstHistory(std::ofstream& histFile) const;
+
+    /// Set gravitational acceleration as a float3 vector.
+    void SetGravitationalAcceleration(const float3 g);
 };
 
 // -----------------------------------------------------------------------------
@@ -302,8 +307,11 @@ class CH_GPU_API ChSystemGpu {
 /// Interface to a Chrono::Gpu mesh system.
 class CH_GPU_API ChSystemGpuMesh : public ChSystemGpu {
   public:
-    /// Construct system with given sphere radius, density, and big domain dimensions.
-    ChSystemGpuMesh(float sphere_rad, float density, float3 boxDims);
+    /// Construct system with given sphere radius, density, big domain dimensions and center.
+    ChSystemGpuMesh(float sphere_rad,
+                    float density,
+                    const ChVector<float>& boxDims,
+                    ChVector<float> O = ChVector<float>(0));
 
     /// Construct system with a checkpoint file.
     ChSystemGpuMesh(const std::string& checkpoint);